create separate controller and model for user features
